# public/openapi/openapi.yaml
openapi: 3.2.0
info:
    title: 'Rotor API'
    version: 1.0.0
    description: |
        Для доступа к данным нужен API-ключ, который можно получить через авторизацию или на странице Мои данные.

        Лимиты запросов: 120 в минуту на все методы, 10 в минуту на авторизацию.
        Остаток лимита — в заголовках X-RateLimit-Limit и X-RateLimit-Remaining, при превышении приходит 429 и Retry-After.
security:
    -
        BearerToken: {  }
components:
    securitySchemes:
        BearerToken:
            type: http
            scheme: bearer
            description: 'Передайте API-ключ в заголовке: Authorization: Bearer {token}'
...
    schemas:
        UserProfile:
            type: object
            properties:
                login:
                    type: string
                name:
                    type: string
                level:
                    type: string
                    enum:
                        - boss
                        - admin
                        - moder
                        - editor
                        - user
                        - pended
                        - banned
                country:
                    type:
                        - string
                        - null
                city:
                    type:
                        - string
                        - null
                info:
                    type:
                        - string
                        - null
                site:
                    type:
                        - string
                        - null
                gender:
                    type:
                        - string
                        - null
                    enum:
                        - male
                        - female
                        - null
                birthday:
                    type:
                        - string
                        - null
                themes:
                    type: string
                    description: 'Текущая тема оформления'
                point:
                    type: integer
                    description: Баллы
                money:
                    type: integer
                    description: Монеты
                status:
                    type:
                        - string
                        - null
                color:
                    type:
                        - string
                        - null
                    description: 'Цвет имени пользователя'
                avatar:
                    type:
                        - string
                        - null
                    format: uri
                picture:
                    type:
                        - string
                        - null
                    format: uri
                rating:
                    type: integer
                language:
                    type: string
                timezone:
                    type: string
                lastlogin:
                    type:
                        - string
                        - null
                    format: date-time
                email:
                    type:
                        - string
                        - null
                    format: email
                phone:
                    type:
                        - string
                        - null
                allprivat:
                    type: integer
                    description: 'Всего приватных сообщений'
                newprivat:
                    type: integer
                    description: 'Непрочитанных приватных сообщений'
...
        File:
            type: object
            properties:
                id:
                    type: integer
                name:
                    type: string
                path:
                    type: string
                    format: uri
                size:
                    type: integer
                extension:
                    type: string
                mime_type:
                    type: string
                is_image:
                    type: boolean
                is_audio:
                    type: boolean
                is_video:
                    type: boolean
...
paths:
...
    /files:
        get:
            tags:
                - Вложения
            summary: 'Свои вложения'
            description: 'Возвращает файлы текущего пользователя. Без id — загруженные, но ещё не привязанные к записи. Требует токен'
            operationId: getFiles
            parameters:
                -
                    name: type
                    in: query
                    required: true
                    schema:
                        type: string
                    description: 'Тип записи (comments, messages, offers, photos, downs)'
                -
                    name: id
                    in: query
                    schema:
                        type: integer
                        default: 0
                    description: 'Id записи; 0 — файлы, ожидающие привязки'
            responses:
                200:
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    data:
                                        type: array
                                        items:
                                            $ref: '#/components/schemas/File'
                400:
                    description: 'Токен не передан'
                401:
                    description: 'Неверный токен'
                422:
                    description: 'Неизвестный тип'
                429:
                    description: 'Превышен общий лимит запросов'
        post:
            tags:
                - Вложения
            summary: 'Загрузка вложения'
            description: |
                Требует токен. Ограничения на размер, количество и расширения приходят в /config.

                id = 0 — запись ещё не создана: файл повиснет за пользователем и привяжется при её сохранении.
            operationId: uploadFile
            requestBody:
                required: true
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            required:
                                - type
                                - file
                            properties:
                                type:
                                    type: string
                                    description: 'Тип записи (comments, messages, offers, photos, downs)'
                                id:
                                    type: integer
                                    default: 0
                                file:
                                    type: string
                                    format: binary
            responses:
                201:
                    description: 'Файл загружен'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message:
                                        type: string
                                    file:
                                        $ref: '#/components/schemas/File'
                400:
                    description: 'Токен не передан'
                401:
                    description: 'Неверный токен'
                422:
                    description: 'Неизвестный тип, превышен лимит файлов, дубликат или файл не прошёл проверку'
                429:
                    description: 'Превышен общий лимит запросов'
    /files/{id}:
        delete:
            tags:
                - Вложения
            summary: 'Удаление вложения'
            description: 'Удалить можно своё вложение (администрация — любое). Требует токен'
            operationId: deleteFile
            parameters:
                -
                    name: id
                    in: path
                    required: true
                    schema:
                        type: integer
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required:
                                - type
                            properties:
                                type:
                                    type: string
            responses:
                200:
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message:
                                        type: string
                400:
                    description: 'Токен не передан'
                401:
                    description: 'Неверный токен'
                422:
                    description: 'Неизвестный тип, файл не найден или принадлежит другому пользователю'
                429:
                    description: 'Превышен общий лимит запросов'
...
    /user:
        get:
            tags:
                - Пользователи
            summary: 'Профиль текущего пользователя'
            operationId: getUser
            responses:
                200:
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    data: { $ref: '#/components/schemas/UserProfile' }
                400:
                    description: 'Api token missing'
                401:
                    description: Unauthorized
...